Added Statuscode and error handler

+++ Controllers/SongsController.cs
@@ -1,4 +1,5 @@
-﻿using Microsoft.AspNetCore.Mvc;
+﻿using Microsoft.AspNetCore.Http;
+using Microsoft.AspNetCore.Mvc;
 using Musicplayer.Data;
 using Musicplayer.Models;
 using System;
@@ -25,44 +26,61 @@ namespace Musicplayer.Controllers
 
         // GET: api/<SongsController>
         [HttpGet]
-        public IEnumerable<Song> Get()
+        public IActionResult Get()
         {
-            return _dbContext.Songs;
+            return Ok(_dbContext.Songs);
         }
         
         // GET api/<SongsController>/5
         [HttpGet("{id}")]
-        public Song Get(int id)
+        public IActionResult Get(int id)
         {
              var song= _dbContext.Songs.Find(id);
-            return song;
+            if (song == null)
+            {
+                return NotFound("NO RECORD FOUND");
+            }
+            return Ok(song);
         }
 
         // POST api/<SongsController>
         [HttpPost]
-        public void Post([FromBody] Song song)
+        public IActionResult Post([FromBody] Song song)
         {
             _dbContext.Songs.Add(song);
              _dbContext.SaveChanges();
+            return StatusCode(StatusCodes.Status201Created);
         }
 
         // PUT api/<SongsController>/5
         [HttpPut("{id}")]
-        public void Put(int id, [FromBody] Song songOb)
+        public IActionResult Put(int id, [FromBody] Song songOb)
         {
             var song = _dbContext.Songs.Find(id);
+
+            if (song == null)
+            {
+                return NotFound("NO RECORD FOUND");
+            }
             song.Title = songOb.Title;
             song.Language = songOb.Language;
             _dbContext.SaveChanges();
+            return Ok("Records updated");
         }
 
         // DELETE api/<SongsController>/5
         [HttpDelete("{id}")]
-        public void Delete(int id)
+        public ActionResult Delete(int id)
         {
             var song = _dbContext.Songs.Find(id);
+
+            if (song == null)
+            {
+                return NotFound("NO RECORD FOUND");
+            }
             _dbContext.Songs.Remove(song);
             _dbContext.SaveChanges();
+            return Ok("Dealted");
         }
     }
 }
